+++ src/main/java/supermarket/employee/crud/DeleteEmployee.java
@@ -30,19 +30,33 @@ public class DeleteEmployee extends HttpServlet {
 	/**
 	 * @see HttpServlet#doPost(HttpServletRequest request, HttpServletResponse response)
 	 */
-	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {  
-        HttpSession session=request.getSession();
-        PrintWriter out = response.getWriter();
-        if(session.getAttribute("role").equals("Admin")) {
-        	int employeeID = Integer.parseInt(request.getParameter("empID"));
-    		Service.deleteEmployee(employeeID);
-    		response.sendRedirect("employee.jsp");
-        }else {
-        	out.println("<script type=\"text/javascript\">");
-			out.println("alert('you dont have permissions');");
-			out.println("location='employee.jsp';");
-			out.println("</script>");
-        }
+	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+		// Retrieve the CSRF token from the request
+		String requestToken = request.getParameter("csrfToken");
+
+// Retrieve the CSRF token from the session
+		String sessionToken = (String) request.getSession().getAttribute("csrfToken");
+
+		if (requestToken != null && requestToken.equals(sessionToken)) {
+			// CSRF token is valid; process the request
+			// ...
+			HttpSession session=request.getSession();
+			PrintWriter out = response.getWriter();
+			if(session.getAttribute("role").equals("Admin")) {
+				int employeeID = Integer.parseInt(request.getParameter("empID"));
+				Service.deleteEmployee(employeeID);
+				response.sendRedirect("employee.jsp");
+			}else {
+				out.println("<script type=\"text/javascript\">");
+				out.println("alert('you dont have permissions');");
+				out.println("location='employee.jsp';");
+				out.println("</script>");
+			}
+		} else {
+			// Invalid CSRF token; handle the error (e.g., return an error page or response)
+			response.setStatus(HttpServletResponse.SC_FORBIDDEN); // You can choose an appropriate status code
+		}
+
 		
 	}
