+++ src/main/java/supermarket/order/crud/DeleteOrder.java
@@ -30,18 +30,33 @@ public class DeleteOrder extends HttpServlet {
 	 * @see HttpServlet#doPost(HttpServletRequest request, HttpServletResponse response)
 	 */
 	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-		HttpSession session=request.getSession();
-        PrintWriter out = response.getWriter();
-        if(session.getAttribute("role").equals("Admin")) {
-        	int orderID = Integer.parseInt(request.getParameter("orderID"));
-    		Service.deleteOrder(orderID);
-    		response.sendRedirect("order.jsp");
-        }else {
-        	out.println("<script type=\"text/javascript\">");
-			out.println("alert('you dont have permissions');");
-			out.println("location='order.jsp';");
-			out.println("</script>");
-        }
+		// Retrieve the CSRF token from the request
+		String requestToken = request.getParameter("csrfToken");
+
+// Retrieve the CSRF token from the session
+		String sessionToken = (String) request.getSession().getAttribute("csrfToken");
+
+		if (requestToken != null && requestToken.equals(sessionToken)) {
+			// CSRF token is valid; process the request
+			// ...
+			HttpSession session=request.getSession();
+			PrintWriter out = response.getWriter();
+			if(session.getAttribute("role").equals("Admin")) {
+				int orderID = Integer.parseInt(request.getParameter("orderID"));
+				Service.deleteOrder(orderID);
+				response.sendRedirect("order.jsp");
+			}else {
+				out.println("<script type=\"text/javascript\">");
+				out.println("alert('you dont have permissions');");
+				out.println("location='order.jsp';");
+				out.println("</script>");
+			}
+		} else {
+			// Invalid CSRF token; handle the error (e.g., return an error page or response)
+			response.setStatus(HttpServletResponse.SC_FORBIDDEN); // You can choose an appropriate status code
+		}
+
+
 		
 	}
